requirements.txt
gevent == 1.2.2
requests == 2.18.1
PyYAML == 3.12
